// src/api-client/models/CreateUpdateEnergyTariff.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { CreateUpdateEnergyTariffElectricity } from "./CreateUpdateEnergyTariffElectricity";
import {
  CreateUpdateEnergyTariffElectricityFromJSON,
  CreateUpdateEnergyTariffElectricityFromJSONTyped,
  CreateUpdateEnergyTariffElectricityToJSON,
} from "./CreateUpdateEnergyTariffElectricity";
import type { CreateUpdateEnergyTariffGas } from "./CreateUpdateEnergyTariffGas";
import {
  CreateUpdateEnergyTariffGasFromJSON,
  CreateUpdateEnergyTariffGasFromJSONTyped,
  CreateUpdateEnergyTariffGasToJSON,
} from "./CreateUpdateEnergyTariffGas";

/**
 *
 * @export
 * @interface CreateUpdateEnergyTariff
 */
export interface CreateUpdateEnergyTariff {
  /**
   * The date in ISO 8601 format (YYYY-MM-DD). This schema is used to represent dates in various contexts, such as event dates, deadlines, or any other date-related information.
   *
   * @type {Date}
   * @memberof CreateUpdateEnergyTariff
   */
  validFrom: Date;
  /**
   *
   * @type {CreateUpdateEnergyTariffElectricity}
   * @memberof CreateUpdateEnergyTariff
   */
  electricity: CreateUpdateEnergyTariffElectricity;
  /**
   *
   * @type {CreateUpdateEnergyTariffGas}
   * @memberof CreateUpdateEnergyTariff
   */
  gas: CreateUpdateEnergyTariffGas;
}

/**
 * Check if a given object implements the CreateUpdateEnergyTariff interface.
 */
export function instanceOfCreateUpdateEnergyTariff(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "validFrom" in value;
  isInstance = isInstance && "electricity" in value;
  isInstance = isInstance && "gas" in value;

  return isInstance;
}

export function CreateUpdateEnergyTariffFromJSON(
  json: any,
): CreateUpdateEnergyTariff {
  return CreateUpdateEnergyTariffFromJSONTyped(json, false);
}

export function CreateUpdateEnergyTariffFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CreateUpdateEnergyTariff {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    validFrom: new Date(json["validFrom"]),
    electricity: CreateUpdateEnergyTariffElectricityFromJSON(
      json["electricity"],
    ),
    gas: CreateUpdateEnergyTariffGasFromJSON(json["gas"]),
  };
}

export function CreateUpdateEnergyTariffToJSON(
  value?: CreateUpdateEnergyTariff | null,
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    validFrom: value.validFrom.toISOString().substring(0, 10),
    electricity: CreateUpdateEnergyTariffElectricityToJSON(value.electricity),
    gas: CreateUpdateEnergyTariffGasToJSON(value.gas),
  };
}